# openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: App API
    version: 0.0.1
paths:
    /api/app_server/fil_usdt:
        get:
            tags:
                - App
            operationId: App_FilUsdt
            parameters:
                - name: symbol
                  in: query
                  schema:
                    type: string
                - name: interval
                  in: query
                  schema:
                    type: string
                - name: limit
                  in: query
                  schema:
                    type: string
                - name: startTime
                  in: query
                  schema:
                    type: string
                - name: endTime
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/FilUsdtReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/app_server/get_per_second_dfil_total:
        get:
            tags:
                - App
            operationId: App_GetPerSecondDFilTotal
            parameters:
                - name: startTime
                  in: query
                  schema:
                    type: integer
                    format: int64
                - name: endTime
                  in: query
                  schema:
                    type: integer
                    format: int64
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetPerSecondDFilTotalReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/app_server/get_per_second_pair_info:
        get:
            tags:
                - App
            operationId: App_GetPerSecondPairInfo
            parameters:
                - name: startTime
                  in: query
                  schema:
                    type: integer
                    format: int64
                - name: endTime
                  in: query
                  schema:
                    type: integer
                    format: int64
                - name: pair
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetPerSecondPairInfoReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/app_server/req_contract:
        get:
            tags:
                - App
            operationId: App_ReqContract
            parameters:
                - name: contractType
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ReqContractReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/app_server/set_per_second_dfil_total:
        get:
            tags:
                - App
            operationId: App_SetPerSecondDFilTotal
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SetPerSecondDFilTotalReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/app_server/set_per_second_pair_info:
        get:
            tags:
                - App
            operationId: App_SetPerSecondPairInfo
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SetPerSecondPairInfoReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
        FilUsdtReply:
            type: object
            properties:
                dataListK:
                    type: array
                    items:
                        $ref: '#/components/schemas/FilUsdtReply_List_k'
        FilUsdtReply_List_k:
            type: object
            properties:
                StartTime:
                    type: integer
                    format: int64
                EndTime:
                    type: integer
                    format: int64
                StartPrice:
                    type: number
                    format: double
                TopPrice:
                    type: number
                    format: double
                LowPrice:
                    type: number
                    format: double
                EndPrice:
                    type: number
                    format: double
                DealTotalAmount:
                    type: number
                    format: double
                DealAmount:
                    type: number
                    format: double
                DealTotal:
                    type: integer
                    format: int64
                DealSelfTotalAmount:
                    type: number
                    format: double
                DealSelfAmount:
                    type: number
                    format: double
        GetPerSecondDFilTotalReply:
            type: object
            properties:
                dataListDFil:
                    type: array
                    items:
                        $ref: '#/components/schemas/GetPerSecondDFilTotalReply_List_dfil'
        GetPerSecondDFilTotalReply_List_dfil:
            type: object
            properties:
                time:
                    type: integer
                    format: int64
                totalSupply:
                    type: string
        GetPerSecondPairInfoReply:
            type: object
            properties:
                dataListPair:
                    type: array
                    items:
                        $ref: '#/components/schemas/GetPerSecondPairInfoReply_List_pair'
        GetPerSecondPairInfoReply_List_pair:
            type: object
            properties:
                time:
                    type: integer
                    format: int64
                reserve0:
                    type: string
                reserve1:
                    type: string
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        ReqContractReply:
            type: object
            properties: {}
        SetPerSecondDFilTotalReply:
            type: object
            properties: {}
        SetPerSecondPairInfoReply:
            type: object
            properties: {}
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
tags:
    - name: App